Enable multi-select filtering for facility types in search
Supports multiple facility types in search criteria by serializing the selection collection. Moves the Select2 multi-select initialization into a shared JavaScript file and ensures the FacilityTypeId is correctly deserialized from the query string during search.

Fixes #996
