+++ FMS/Pages/Facilities/Index.cshtml
@@ -11,6 +11,7 @@
     <script src="~/lib/select2/js/select2.min.js"></script>
     <script src="~/js/formSearch.js"></script>
     <script src="~/js/formIndex.js"></script>
+    <script src="~/js/MultiSelectFacType.js"></script>
     <script>
         $( '#select-field' ).select2( {
             theme: 'bootstrap-5'
