+++ FMS/Pages/Facilities/Map.cshtml
@@ -18,6 +18,7 @@
     <script src="~/lib/OverlappingMarkerSpiderfier/oms.min.js"></script>
     <script src="~/lib/select2/js/select2.min.js"></script>
     <script src="~/js/mapSearch.js"></script>
+    <script src="~/js/MultiSelectFacType.js"></script>
     <script>
         $( '#select-field' ).select2( {
             theme: 'bootstrap-5'
